Create link in index and lasagna html files
Created a link in index.html that leads to lasagna.html.The link allows the users to navigate to the
lasagna page easily. I also created a link in lasagna.html that leads the user to the index page. This
makes it easier to get back to the main page without having to press the left arrow in the top left.